Add files via upload
These are the pages for the userid and  password recovery.  forgotpassword.php and resetpassword.php. I tried the change password and kept getting stuck. 
This link will need added to the end of the login.php form. 
Forgotten your <a href="forgotpassword.php">user id or password</a>?
For the reset password to work I had to add another colomn to the quizusers table called token.  So when the quizusers table is created this column will need to be added.